+++ ts/src/program/namespace/transaction.ts
@@ -16,8 +16,13 @@ export default class TransactionFactory {
     const txFn: TransactionFn<IDL, I> = (...args): Transaction => {
       const [, ctx] = splitArgsAndCtx(idlIx, [...args]);
       const tx = new Transaction();
+      if (ctx.preInstructions && ctx.instructions) {
+        throw new Error("instructions is deprecated, use preInstructions");
+      }
+      ctx.preInstructions?.forEach((ix) => tx.add(ix));
       ctx.instructions?.forEach((ix) => tx.add(ix));
       tx.add(ixFn(...args));
+      ctx.postInstructions?.forEach((ix) => tx.add(ix));
       return tx;
     };